Completely separate additions and deletions Before, the SPO-sorted additions were stored together with the deletions in the single SPO-sorted deletions tree. This worked fine, except for the case when only additions were present and lookups had to be done. These lookups require several deletion_counts, which requires looking at the last deletion in the tree and getting its PatchPosition. When there are no deletions, this deletion_count is extremely slow, since it has to go every single element in the tree until no more are available.
This has been solved by moving all additions to a new SPO tree.
This required a lot of refactoring all over the place.
It was now possible to simplify some things, like for example the calculation of local changes in patches, which now happens automatically, and no separate step is needed.